backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
firebase-admin==6.4.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AI & ML
google-generativeai==0.3.2
spacy==3.7.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
sentence-transformers==2.3.1

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Google Cloud
google-cloud-speech==2.24.0
google-cloud-texttospeech==2.16.1
google-cloud-storage==2.14.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
